feat(07): complete chapter on multi-threading
This topic is entirely foreign to me, I learnt programming with PHP. The
exercises were manageable enough, but I doubt I'd be able to apply all
this in the real world just like that. It's exciting the way it
introduces a whole new conceptual area to reason about when writing
programs. Multithreading can accomplish way more throughput, but only if
the threads aren't constantly blocking waiting for locks. You have to
find areas in the problem domain where you can divide work, carving at
the joint if you will.

From my understanding, you cannot have simultaneous read and write
access across threads, that would be a data race and Rust makes it a
type error. Thinking back to the simpler example of multithreading:
calculating the sum or product of a list; you **can** break work into
chunks where each thread can contribute a part of the work that later
has to be combined. But to do that, each thread needs their own copy of
a chunk of the list of numbers.